requirements.txt
asgiref==3.2.10
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.3
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.1.2
django-crispy-forms==1.9.2
django-datetime-widget==0.9.3
django-session-security==2.6.6
django-session-timeout==0.1.0
django-weasyprint==1.0.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
et-xmlfile==1.0.1
html5lib==1.1
jdcal==1.4.1
mysqlclient==2.0.1
openpyxl==3.0.5
paho-mqtt==1.5.1
Pillow==7.2.0
pycparser==2.20
PyJWT==1.7.1
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
WeasyPrint==51
webencodings==0.5.1
